add imitationDesired function

+++ ohrc_control/include/ohrc_control/ohrc_controller.hpp
@@ -126,7 +126,7 @@ protected:
   // virtual std::shared_ptr<Interface> selectInterface(std::shared_ptr<CartController> cartController) {
   // }
 
-  int getNRobot() {
+  size_t getNRobot() {
     return nRobot;
   }
